ISSUE-11: renamed SupermarketCatalog to TemplateCatalog. also renamed related files.
