+++ pythonsupermarket/pythonsupermarket/template_catalog.py
