+++ pythonsupermarket/tests/test_template_catalog.py
@@ -1,6 +1,6 @@
 import unittest
 
-from pythonsupermarket.catalog import TemplateCatalog
+from pythonsupermarket.template_catalog import TemplateCatalog
 
 
 class TestSupermarketCatalog(unittest.TestCase):
